add API - getLatestNews.
Добавил новый endpoint в api, чтобы получать не только новости по нашим фильтрам, но и просто последние новости и добавил их вывод в баннеры.
