+++ src/components/LatestNews/LatestNews.jsx
@@ -1,11 +1,15 @@
+import { getLatestNews } from "../../api/apiNews";
 import BannersList from "../../components/BannersList/BannersList";
+import { useFetch } from "../../helpers/hooks/useFetch";
 
 import styles from "./styles.module.css";
 
-const LatestNews = ({ banners, isLoading }) => {
+const LatestNews = () => {
+	const { data, isLoading } = useFetch(getLatestNews);
+
 	return (
 		<section className={styles.section}>
-			<BannersList banners={banners} isLoading={isLoading} />
+			<BannersList banners={data && data?.news} isLoading={isLoading} />
 		</section>
 	);
 };
