+++ src/components/NewsByFilters/NewsByFilters.jsx
@@ -1,10 +1,28 @@
-import { TOTAL_PAGES } from "../../constants/constants";
+import { getNews } from "../../api/apiNews";
+import { PAGE_SIZE, TOTAL_PAGES } from "../../constants/constants";
+import { useDebounce } from "../../helpers/hooks/useDebounce";
+import { useFetch } from "../../helpers/hooks/useFetch";
+import { useFilters } from "../../helpers/hooks/useFilters";
 import NewsFilters from "../NewsFilters/NewsFilters";
 import NewsList from "../NewsList/NewsList";
-import Pagination from "../Pagination/Pagination";
+import PaginationWrapper from "../PaginationWrapper/PaginationWrapper";
 import styles from "./styles.module.css";
 
-const NewsByFilters = ({ filters, changeFilters, isLoading, news }) => {
+const NewsByFilters = () => {
+	const { filters, changeFilters } = useFilters({
+		page_number: 1,
+		page_size: PAGE_SIZE,
+		category: "all",
+		keywords: "",
+	});
+
+	const debouncedKeywords = useDebounce(filters.keywords, 1000);
+
+	const { data, isLoading } = useFetch(getNews, {
+		...filters,
+		keywords: debouncedKeywords,
+	});
+
 	const handlePreviousPage = () => {
 		if (filters.page_number > 1) {
 			changeFilters("page_number", filters.page_number - 1);
@@ -25,25 +43,18 @@ const NewsByFilters = ({ filters, changeFilters, isLoading, news }) => {
 		<section className={styles.section}>
 			<NewsFilters filters={filters} changeFilters={changeFilters} />
 
-			<Pagination
+			<PaginationWrapper
+				top
+				bottom
 				totalPages={TOTAL_PAGES}
 				handleNextPage={handleNextPage}
 				handlePreviousPage={handlePreviousPage}
 				handlePageClick={handlePageClick}
 				currentPage={filters.page_number}
 				isLoading={isLoading}
-			/>
-
-			<NewsList isLoading={isLoading} news={news} />
-
-			<Pagination
-				totalPages={TOTAL_PAGES}
-				handleNextPage={handleNextPage}
-				handlePreviousPage={handlePreviousPage}
-				handlePageClick={handlePageClick}
-				currentPage={filters.page_number}
-				isLoading={isLoading}
-			/>
+			>
+				<NewsList isLoading={isLoading} news={data?.news} />
+			</PaginationWrapper>
 		</section>
 	);
 };
