+++ src/components/PaginationWrapper/PaginationWrapper.jsx
@@ -1,0 +1,13 @@
+import Pagination from "../Pagination/Pagination";
+
+const PaginationWrapper = ({ top, bottom, children, ...paginationProps }) => {
+	return (
+		<>
+			{top && <Pagination {...paginationProps} />}
+			{children}
+			{bottom && <Pagination {...paginationProps} />}
+		</>
+	);
+};
+
+export default PaginationWrapper;
