+++ src/helpers/hooks/useFetch.js
@@ -1,4 +1,3 @@
-import axios from "axios";
 import { useEffect, useState } from "react";
 
 const useFetch = (fetchFunction, params = null) => {
